add tool palette

+++ src/Impl/Util/Sfml.h
@@ -25,7 +25,7 @@ plug::Texture getPlugTexture(const char *texture_path);
 
 plug::Texture getPlugTexture(const sf::Texture &texture);
 
-plug::Color HSV2RGB(uint16_t i_hue, uint8_t sat, uint8_t val);
+plug::Color HSV2RGB(size_t i_hue, size_t sat, size_t val);
 
 bool checkColors(plug::Color rhs, plug::Color lhs);
